+++ templates/master/default-settings.js
@@ -22,12 +22,16 @@ var default_settings = {
     ALT_SEARCH_KENDRA_TOP_ANSWER_MESSAGE: "Amazon Kendra suggested answer.",
     ALT_SEARCH_KENDRA_FAQ_MESSAGE: "Answer from Amazon Kendra FAQ.",
     ALT_SEARCH_KENDRA_ANSWER_MESSAGE: "While I did not find an exact answer, these search results from Amazon Kendra might be helpful.", //Message displayed when a search comes from Kemdra
+    ALT_SEARCH_KENDRA_RESPONSE_TYPES: "ANSWER,DOCUMENT,QUESTION_ANSWER",
+    ALT_SEARCH_KENDRA_ABBREVIATE_MESSAGE_FOR_SSML: "true",
     KENDRA_FAQ_INDEX: "${DefaultKendraIndexId}", // Kendra Index specific for FAQ for if Kendra FAQ sync is enabled
     KENDRA_FAQ_CONFIG_MAX_RETRIES: 8,    //User can override number of max retries in AWS SDK configurations
     KENDRA_FAQ_CONFIG_RETRY_DELAY: 600,  //User can override number of miliseconds delay between retries in AWS SDK configurations
     KENDRA_FAQ_ES_FALLBACK: "true", // Optional ElasticSearch Fallback engine for if KendraFAQ fails
     ENABLE_KENDRA_WEB_INDEXER:"false", //Enables web crawler -- indexes pages specified by KENDRA_INDEXER_URLS
     KENDRA_INDEXER_URLS: "",//comma separated list of urls for Kendra to crawler
+    KENDRA_INDEXER_CRAWL_DEPTH: 3,
+    KENDRA_INDEXER_CRAWL_MODE: "SUBDOMAINS", // Should be one of 'HOST_ONLY'|'SUBDOMAINS'|'EVERYTHING'
     KENDRA_INDEXER_SCHEDULE: "rate(1 day)",//See https://docs.aws.amazon.com/AmazonCloudWatch/latest/events/ScheduledEvents.html for valid expressions
     KENDRA_WEB_PAGE_INDEX: "${DefaultKendraIndexId}",//The index to use for the web crawler, a custom data source will automatically be added to the specified index.  The index will automatically be added to ALT_SEARCH_KENDRA_INDEXES
     ERRORMESSAGE: "Unfortunately I encountered an error when searching for your answer. Please ask me again later.",
@@ -54,13 +58,16 @@ var default_settings = {
     PII_REJECTION_QUESTION: "pii_rejection_question", // If PII is found, the user's request (question) will change to this phrase
     PII_REJECTION_WITH_COMPREHEND: true, //When set to true, Comprehend will be used for PII detection in addition to matching  based on the PII_REJECTION_REGE
     PII_REJECTION_REGEX: "\\b\\d{4}\\b(?![-])|\\b\\d{9}\\b|\\b\\d{3}-\\d{2}-\\d{4}\\b", //Regex to use to find PII.,
-    PII_REJECTION_IGNORE_TYPES: "Name,Address", //See https://aws.amazon.com/blogs/machine-learning/detecting-and-redacting-pii-using-amazon-comprehend/ for valid types
+    PII_REJECTION_ENTITY_TYPES:"ADDRESS,EMAIL,SSN,PHONE,PASSWORD,BANK_ACCOUNT_NUMBER,BANK_ROUTING,CREDIT_DEBIT_NUMBER", //See https://aws.amazon.com/blogs/machine-learning/detecting-and-redacting-pii-using-amazon-comprehend/ for valid types
+    PII_REJECTION_CONFIDENCE_SCORE: .99,
     DISABLE_CLOUDWATCH_LOGGING: "false", // disable all logging in fulfillment es query handler lambda. does not disable logging from Lambda Hooks or Conditional Chaining Lambda functions
     MINIMAL_ES_LOGGING: "false", // do not log utterances or session attributes to elasticsearch for kibana logging
     S3_PUT_REQUEST_ENCRYPTION: "", // enable header x-amz-server-side-encryption header and set with this value
     BOT_ROUTER_WELCOME_BACK_MSG: "Welcome back to QnABot.", // The text used by QnABot when ending communication from a specialty bot
     BOT_ROUTER_EXIT_MSGS: "exit,quit,goodbye,leave", // The exit phrases in comma separated list available for the a user to end communication with a specialty bot
     RUN_LAMBDAHOOK_FROM_QUERY_STEP: "true",
+    LAMBDA_PREPROCESS_HOOK: "",
+    LAMBDA_POSTPROCESS_HOOK: ""
 };
 
 module.exports = {